+++ .gitignore
@@ -5,3 +5,4 @@ inst/doc
 test.R*
 deploy.R*
 CRAN-SUBMISSION
+.vscode
