--- CRAN-RELEASE
@@ -1,2 +1,0 @@
-This package was submitted to CRAN on 2021-05-04.
-Once it is accepted, delete this file and tag the release (commit 6664429).
