+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: FKF.SP
 Title: Fast Kalman Filtering Through Sequential Processing
-Version: 0.3.2
+Version: 0.3.3
 Authors@R: c(person("Thomas", "Aspinall", email = "tomaspinall2512@gmail.com", 
          role = c("aut", "cre"),
          comment = c(ORCID = "0000-0002-6968-1989")),
